Bug Fix: CSS loading issue
Fixed the bug in the web server that was causing CSS to not
load properly in the web browser. The server was creating
responses that were all Content-Type: text/html. .find(<text>)
does not work as I thought, so changing the order of
operations in which I check file extensions to have html
as the last check fixes this issue.